% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_min_score.R
\name{calculate_score}
\alias{calculate_score}
\title{Find the optimal min_score threshold for \code{searchSeq()}}
\usage{
calculate_score(peaks, pwm, param, genome_build)
}
\arguments{
\item{pwm}{An object of class \code{PWMatrix}.}

\item{genome_build}{The genome build that the peak sequences should be
derived from.}
}
\description{
\code{optimal_min_score()} finds the optimal min_score threshold for
TFBSTools::searchSeq by maximising the difference between the proportion
of motifs found in a subset of the peak data and a set of background control
sequences.
}
\keyword{internal}
