% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_control_input.R
\name{check_control_input}
\alias{check_control_input}
\title{Helper function to check control input}
\usage{
check_control_input(control_input, shuffle_k, peak_sequences, genome_build)
}
\description{
\code{check_control_input} confirms that the control option has been set
correctly. The function returns a DNAStringSet object of control sequences.
}
\keyword{internal}
