% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_peak_input.R
\name{check_peak_input}
\alias{check_peak_input}
\title{Check that the peak input is valid}
\usage{
check_peak_input(peak_input, genome_build)
}
\arguments{
\item{peak_input}{Either a path to the narrowPeak file or a GRanges peak
object generated by \code{read_peak_file()}.}

\item{genome_build}{The genome build that the peak sequences should be
derived from.}
}
\value{
A list containing a GRanges peak object and a DNAStringSet sequence
object.
}
\description{
\code{check_peak_input} confirms that the peak input is valid.
}
\keyword{internal}
