% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creb_motif.R
\docType{data}
\name{creb_motif}
\alias{creb_motif}
\title{CREB Motif}
\format{
A universalmotif object produced by the \code{read_motif_file}
function.
}
\source{
JASPAR2024 \link{https://jaspar.elixir.no/matrix/MA0018.5/}
}
\usage{
creb_motif
}
\description{
This dataset contains the MA0018.5 CREB1 motif from JASPAR2024.
}
\keyword{datasets}
