% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creb_peaks.R
\docType{data}
\name{creb_peaks}
\alias{creb_peaks}
\title{CREB Peaks}
\format{
A GRanges peak object outputted by the \code{read_peak_file}
function.
}
\usage{
creb_peaks
}
\description{
This dataset contains a set of CREB1 TIP-seq peaks (narrowPeak) produced by
MACS3 (first-mate, 5' shift). We have subset the peaks to reduce the data
size (only chr19). The commands used to subset the data were:
}
\details{
creb_peaks <- read.table("path/to/creb_peaks")
creb_peaks <- creb_peaks[creb_peaks$V1 == "chr19",]
write.table(creb_peaks,
            "creb_subset.narrowPeak",
            row.names = FALSE,
            col.names = FALSE,
            quote = FALSE,
            sep = "\t")
}
\keyword{datasets}
