% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctcf_motif.R
\docType{data}
\name{ctcf_motif}
\alias{ctcf_motif}
\title{CTCF Motif}
\format{
A universalmotif object produced by the \code{read_motif_file}
function.
}
\source{
JASPAR2024 \url{https://jaspar.elixir.no/matrix/MA1930.2/}
}
\usage{
data("ctcf_motif")
}
\description{
The MA1930.2 CTCF motif from JASPAR2024.
}
\keyword{datasets}
