% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{Plot the positional distribution of motifs relative to peak summits}
\usage{
density_plot(distance_vec, plot_title = NULL, limits = c(-400, 400))
}
\description{
\code{density_plot} creates a density plot of the positional distribution
of motifs within peaks.
}
