% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_background_model.R
\name{markov_background_model}
\alias{markov_background_model}
\title{Generate a 0-order Markov background model}
\usage{
markov_background_model(sequences, out_dir)
}
\arguments{
\item{sequences}{A DNAStringSet object.}

\item{out_dir}{Location to save the 0-order background file.}
}
\value{
The path to the 0-order background file.
}
\description{
\code{markov_background_model()} generates a 0-order background model for use
with FIMO or AME. The function uses the letter frequencies in the input
sequences to generate the background model.
}
\keyword{internal}
