% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_min_score.R
\name{optimal_min_score}
\alias{optimal_min_score}
\title{Find the optimal min_score for searchSeq()}
\usage{
optimal_min_score(peaks, pwm, genome_build)
}
\arguments{
\item{peaks}{A GRanges peak object}

\item{pwm}{An object of class PWMatrix.}

\item{seed}{A single value specifying the state for random number generation.}
}
\description{
\code{optimal_min_score()} finds the optimal min_score threshold for
TFBSTools::searchSeq by maximising the difference between the proportion
of motifs found in a subset of the peak data and a set of background control
sequences.
}
