% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_proportion.R
\name{peak_proportion}
\alias{peak_proportion}
\title{Calculate the proportion of peaks with a motif}
\usage{
peak_proportion(peak_file, pwm, plot = FALSE, min_score = 0.8)
}
\arguments{
\item{peak_file}{Character string specifying the path to the narrowPeak file.}

\item{pwm}{An object of class PWMatrix.}

\item{plot}{Logical specifying whether or not to create a plot. Note that the
peak file must contain a column suitable for ranking.}

\item{min_score}{Numeric specifying the minimum score a sequence must pass
to be considered a motif. The default is 0.8.}

\item{out_dir}{Path to the output directory where the filtered peak file
will be written.}
}
\description{
\code{peak_proportion()} calculates the proportion of peaks containing a
motif. To generate a plot the peak file must contain a column that can be
used for ranking (e.g. p-values, q-values or signal values).
}
\examples{
\dontrun{
pwm <-
  read_motif_file(motif_file = "./MA0018.3.meme", file_format = "meme")

peak_proportion(peak_file = "SK032_1_SK035_1_SK035_11_R1_peaks.narrowPeak",
                pwm = pwm,
                plot = TRUE,
                min_score = 0.8
                )
                }

}
