% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_and_genes.R
\docType{data}
\name{peaks_and_genes}
\alias{peaks_and_genes}
\title{Find regions of the genome with high numbers of true peaks}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
peaks_and_genes
}
\description{
Slide a window along each chromosome. Within each window we calculate the
number of binding sites and the number of genes. We calculate correlation
between peaks and genes.
}
\keyword{datasets}
