% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_peak_file_seacr.R
\name{read_peak_file_seacr}
\alias{read_peak_file_seacr}
\title{Import a SEACR BED file as a GRanges object}
\usage{
read_peak_file_seacr(file_path)
}
\arguments{
\item{file_path}{Path to a \code{BED} file.}
}
\description{
\code{read_peak_file()} reads a SEACR BED file as a GRanges object and adds
an additional metadata column representing the position of the peak summit.
}
\examples{
\dontrun{
peak_file <- system.file("extdata",
                         "ctcf_seacr.stringent.bed",
                         package = "MotifStats"
                         )
peak_obj <- read_peak_file_seacr(file_path = peak_file)
}

}
