% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summit_to_motif.R
\name{summit_to_motif}
\alias{summit_to_motif}
\title{Calculate the distance between peak summits and motifs}
\usage{
summit_to_motif(peak_file, pwm, min_score = 0.8)
}
\arguments{
\item{peak_file}{Character string specifying the path to the narrowPeak file.}

\item{pwm}{An object of class PWMatrix.}

\item{min_score}{Numeric specifying the minimum score a sequence must pass
to be considered a motif. The default is 0.8.}
}
\value{
GRanges peak object with metadata columns relating to motif
positions.
}
\description{
\code{summit_to_motif()} calculates the distance between each motif and its
nearest summit. A peak summit represents the exact nucleotide where signal
enrichment is highest.
}
\examples{
\dontrun{
pwm <-
  read_motif_file(motif_file = "./MA0018.3.meme", file_format = "meme")

summit_to_motif(peak_file = "read1_01_peaks.narrowPeak",
                pwm = pwm,
                min_score = 0.8
                )
                }
}
