% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacent_sequences.R
\name{adjacent_sequences}
\alias{adjacent_sequences}
\title{Generate sequences immediately adjacent to peaks}
\usage{
adjacent_sequences(peaks, genome_build)
}
\arguments{
\item{peaks}{A GRanges peak object}
}
\description{
\code{adjacent_sequences()} generates sequences adjacent to each peak in
"peaks".
}
\details{
Adjacent sequences will have the same width as their respective
peaks. Where a peak is at the boundary of a chromosome, the adjacent
sequence width will not extend beyond the boundaries of the chromosome.
}
\keyword{internal}
