% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacent_sequences.R
\name{adjacent_sequences}
\alias{adjacent_sequences}
\title{Generate sequences immediately adjacent to peaks}
\usage{
adjacent_sequences(peaks)
}
\arguments{
\item{peaks}{A GRanges peak object}
}
\description{
\code{adjacent_sequences()} generates sequences adjacent to each peak in
"peaks". These sequences will be used as a reference when calculating the
proportion of peaks with a motif.
}
\keyword{internal}
