% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_peak_input.R
\name{check_peak_input}
\alias{check_peak_input}
\title{Helper function to check peak input}
\usage{
check_peak_input(peak_input, genome_build)
}
\description{
\code{check_peak_input} confirms that the peak input is valid and returns a
DNAStringSet object.
}
\keyword{internal}
