% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creb_motif.R
\docType{data}
\name{creb_motif}
\alias{creb_motif}
\title{CREB Motif}
\format{
A universalmotif object produced by the \code{read_motif_file}
function.
}
\source{
JASPAR2024 \url{https://jaspar.elixir.no/matrix/MA0018.5/}
}
\usage{
data("creb_motif")
}
\description{
The MA0018.5 CREB1 motif from JASPAR2024.
}
\keyword{datasets}
