% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creb_peaks.R
\docType{data}
\name{creb_peaks}
\alias{creb_peaks}
\title{CREB Peaks}
\format{
A GRanges peak object outputted by the \code{read_peak_file}
function.
}
\usage{
creb_peaks
}
\description{
This dataset contains a set of CREB1 TIP-seq peaks (narrowPeak) produced by
MACS3.
}
\keyword{datasets}
