% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creb_peaks.R
\docType{data}
\name{creb_peaks}
\alias{creb_peaks}
\title{CREB Peaks}
\format{
A GRanges peak object outputted by the \code{read_peak_file}
function.
}
\usage{
data("creb_peaks")
}
\description{
A set of CREB1 TIP-seq peaks (narrowPeak) produced by MACS3 (first-mate only,
5'-shift to centre the 5' end of each read). The BAM file used as input for
the peak calling is not available publicly. The peak file has been subset
to reduce the data size (only chromosome 19).
}
\details{
The code used to prepare the data as an .rda file is:

\code{creb_peaks <- read.table("path/to/creb_peaks.narrowPeak")}
\code{creb_peaks <- creb_peaks[creb_peaks$V1 == "chr19",]}
\code{write.table(creb_peaks,
                  "creb_subset.narrowPeak",
                  row.names = FALSE,
                  col.names = FALSE,
                  quote = FALSE,
                  sep = "\t")}
\code{creb_peaks <- read_peak_file("inst/extdata/creb_subset.narrowPeak")}
\code{usethis::use_data(creb_peaks, overwrite = TRUE)}
}
\keyword{datasets}
