% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{Plot the positional distribution of motifs relative to peak summits}
\usage{
density_plot(
  distance_vec,
  plot_title = NULL,
  x_label = NULL,
  y_label = NULL,
  limits = c(-400, 400)
)
}
\arguments{
\item{distance_vec}{A vector of summit-to-motif distances. This is an
output of the \code{summit_to_motif} function.}

\item{plot_title}{Title for the density plot.}

\item{x_label}{X-axis label.}

\item{y_label}{Y-axis label.}

\item{limits}{A vector of length 2 denoting the lower and upper bounds of the
x-axis. The default is \code{c(-400, 400)}.}
}
\description{
\code{density_plot} creates a density plot of the positional distribution
of motifs within peaks.
}
