% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_motif_hits.R
\name{get_motif_hits}
\alias{get_motif_hits}
\title{Scan a set of peaks for a specified motif}
\usage{
get_motif_hits(peaks, pwm, min_score)
}
\arguments{
\item{peaks}{A GRanges peak object.}

\item{pwm}{An object of class PWMatrix.}

\item{min_score}{Numeric specifying the minimum score a sequence must pass
to be considered a motif. The default is 0.8.}
}
\description{
\code{get_motif_hits()} uses the \code{searchSeq()} function from
\code{TFBSTools} to scan each peak for the presence of a motif.
}
\keyword{internal}
