% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_proportion.R
\name{peak_proportion}
\alias{peak_proportion}
\title{Calculate the proportion of peaks with a motif}
\usage{
peak_proportion(peak_input, motif, fp_rate = 0.05, genome_build)
}
\arguments{
\item{peak_input}{Either a path to the narrowPeak file or a GRanges peak
object generated by \code{read_peak_file()}.}

\item{fp_rate}{False-positive rate to be used to calculate an appropriate
p-value threshold,}

\item{genome_build}{The genome build that the peak sequences should be
derived from.}

\item{pwm}{An object of class \code{PWMatrix}.}
}
\value{
A list containing a FIMO results data frame and a numeric referring
to the proportion of peaks with a motif.
}
\description{
\code{peak_proportion()} calculates the proportion of peaks containing a
motif relative to a set of background sequences using Find Individual Motif
Occurrences (FIMO) from \link[memes].
}
\examples{
\dontrun{
peak_file <- system.file("extdata",
                         "rep1_peaks.narrowPeak",
                         package = "MotifStats"
                         )
data("creb_motif", package = "MotifStats")

peak_proportion(
  peak_input = peak_file,
  motif = creb_motif,
  fp_rate = 5e-02,
  genome_build = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
)
}

}
\seealso{
\link[memes]{runFimo}
}
