% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_proportion.R
\name{peak_proportion}
\alias{peak_proportion}
\title{Calculate the proportion of peaks with a motif}
\usage{
peak_proportion(
  peak_input,
  control_input = "shuffle",
  pwm,
  genome_build,
  shuffle_k = 2,
  min_score = 0.8,
  optimal_min_score = FALSE,
  num_samples = 4,
  sample_size = 250,
  num_threads = 1L,
  deoptim_lower = 0.5,
  deoptim_upper = 1,
  deoptim_np = 20,
  deoptim_itermax = 25,
  deoptim_strategy = 6,
  seed = NULL
)
}
\arguments{
\item{peak_input}{Either a path to the narrowPeak file or a GRanges peak
object generated by \code{read_peak_file()}.}

\item{control_input}{Method for generating control sequences.
\code{"shuffle"} will shuffle the input sequences. \code{"adjacent"} will
select real sequences immediately adjacent to each peak. The default is
\code{"shuffle"}.}

\item{pwm}{An object of class \code{PWMatrix}.}

\item{genome_build}{The genome build that the peak sequences should be
derived from.}

\item{shuffle_k}{K-let size. This option is only used if \code{control_input}
is set to \code{"shuffle"}.}

\item{min_score}{Numeric specifying the minimum score a sequence must pass
to be considered a motif.}

\item{optimal_min_score}{If TRUE, numerical optimisation is used to find an
optimal min_score parameter for the motif search. We use the differential
evolution algorithm implemented in deoptim.}

\item{num_samples}{Number of samples to use to calculate the optimal
min_score threshold. This option is only used if optimal_min_score is set to
TRUE.}

\item{sample_size}{Number of peaks in each sample. This option is only used
if optimal_min_score is set to TRUE.}

\item{num_threads}{Number of threads to use for min_score optimisation. We
recommend setting this equal to the number of samples for the greatest speed
up.}

\item{deoptim_lower}{The lower bound of the parameter search space.}

\item{deoptim_upper}{The upper bound of the parameter search space.}

\item{deoptim_np}{The number of population members.}

\item{deoptim_itermax}{The number of population generations the optimisation
procedure will use.}

\item{deoptim_strategy}{The differential evolution strategy the optimisation
procedure will use.}

\item{seed}{Set this to an integer if you want the same random background
sequences across function calls (given the same peak_input).}
}
\description{
\code{peak_proportion()} calculates the proportion of peaks containing a
motif relative to a set of background sequences.
}
\details{
When scanning sequences for the presence of a given motif, a difficult task
is selecting a suitable threshold for a "hit". This process typically
involves using a position weight matrix (PWM) to score each sequence based on
its similarity to a motif of interest. A score threshold representing the
proportion of the PWM's maximum possible score is then applied to filter for
significant matches.

The peak_proportion() function uses \code{TFBSTools::searchSeq()} to find
motif occurrences. Similar to above, \code{searchSeq} uses
a \code{min.score} threshold to filter significant hits. We have found that
the optimal \code{min.score} threshold is highly sensitive to the particular
motif you are scanning. To make the process of selecting a score threshold
more seamless, we supply the option to use differential evolution
optimisation to find the score threshold that minimises
Loss = Proportion(background sequences with motif) - Proportion(peak
sequences with motif).
}
\examples{
\dontrun{
peak_file <- system.file("extdata",
                         "rep1_peaks.narrowPeak",
                         package = "MotifStats"
                         )
data("creb_motif", package = "MotifStats")

peak_proportion(
  peak_file = peak_file,
  pwm = creb_motif,
  genome_build = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  min_score = NULL,
  optimal_min_score = TRUE,
  num_samples = 4,
  sample_size = 250,
  num_threads = 4L,
  deoptim_lower = 0.5,
  deoptim_upper = 1,
  deoptim_np = 20,
  deoptim_itermax = 25,
  deoptim_strategy = 6,
  seed = 123 # set seed for reproducibility
)
}

}
\seealso{
\link[DEoptim]{DEoptim}, \link[TFBSTools]{searchSeq}
}
