% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_proportion.R
\name{peak_proportion}
\alias{peak_proportion}
\title{Calculate the proportion of peaks with a motif}
\usage{
peak_proportion(
  peak_input,
  pwm,
  genome_build,
  min_score = 0.8,
  optimal_min_score = FALSE,
  seed = NULL
)
}
\arguments{
\item{peak_input}{Either a path to the narrowPeak file or a GRanges peak
object generated by \code{read_peak_file()}.}

\item{pwm}{An object of class PWMatrix.}

\item{min_score}{Numeric specifying the minimum score a sequence must pass
to be considered a motif.}

\item{optimal_min_score}{If TRUE, numerical optimisation is used to find an
optimal min_score parameter for the motif search}

\item{seed}{Set this to an integer if you want the same random background
sequences across function calls (given the same peak_input). This parameter
is ignored when optimal_min_score is \code{FALSE}.}
}
\description{
\code{peak_proportion()} calculates the proportion of peaks containing a
motif. To generate a plot the peak file must contain a column that can be
used for ranking (e.g. p-values, q-values or signal values).
}
\examples{
\dontrun{

peak_file <- system.file("extdata",
                         "rep1_peaks.narrowPeak",
                         package = "MotifStats"
                         )
data("creb_motif", package = "MotifStats")

peak_proportion(
  peak_file = peak_file,
  pwm = creb_motif,
  genome_build = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  min_score = NULL,
  optimal_min_score = TRUE,
  seed = 123 # set seed for reproducibility
)
}

}
