% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_peak_file.R
\name{read_peak_file}
\alias{read_peak_file}
\title{Import a narrowPeak file as a GRanges object}
\usage{
read_peak_file(file_path)
}
\arguments{
\item{file_path}{Path to a \code{narrowPeak} file.}
}
\description{
\code{read_peak_file()} reads a narrowPeak file as a GRanges object and adds
an additional metadata column representing the position of the peak summit.
}
\examples{
\dontrun{
peak_file <- system.file("extdata",
                         "rep1_peaks.narrowPeak",
                         package = "MotifStats"
                         )
peak_obj <- read_peak_file(file_path = peak_file)
}

}
