% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summit_to_motif.R
\name{summit_to_motif}
\alias{summit_to_motif}
\title{Calculate the distance between peak summits and motifs}
\usage{
summit_to_motif(peak_input, motif, fp_rate = 0.05, genome_build)
}
\arguments{
\item{peak_input}{Either a path to the narrowPeak file or a GRanges peak
object generated by \code{read_peak_file()}.}

\item{fp_rate}{False-positive rate to be used to calculate an appropriate
p-value threshold,}

\item{genome_build}{The genome build that the peak sequences should be
derived from.}
}
\value{
A list containing an expanded GRanges peak object with metadata
columns relating to motif positions along with a vector of summit-to-motif
distances for each valid peak.
}
\description{
\code{summit_to_motif()} calculates the distance between each motif and its
nearest peak summit.
}
\details{
This function is designed to work with MACS2/3 narrowPeak files.
}
\examples{
\dontrun{
pwm <-
  read_motif_file(motif_file = "./MA0018.3.meme", file_format = "meme")

summit_to_motif(
  peak_input = peak_file,
  pwm = creb_motif,
  fp_rate = 5e-02,
  genome_build = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
)
}
}
