% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summit_to_motif.R
\name{summit_to_motif}
\alias{summit_to_motif}
\title{Calculate the distance between peak summits and motifs}
\usage{
summit_to_motif(
  peak_input,
  pwm,
  min_score = 0.8,
  genome_build,
  optimal_min_score = FALSE,
  seed = NULL
)
}
\arguments{
\item{peak_input}{Either a path to the narrowPeak file or a GRanges peak
object generated by \code{read_peak_file()}.}

\item{pwm}{An object of class PWMatrix.}

\item{min_score}{Numeric specifying the minimum score a sequence must pass
to be considered a motif.}

\item{optimal_min_score}{If TRUE, numerical optimisation is used to find an
optimal min_score parameter for the motif search}

\item{seed}{Set this to an integer if you want the same random background
sequences across function calls (given the same peak_input). This parameter
is ignored when optimal_min_score is \code{FALSE}.}
}
\value{
GRanges peak object with metadata columns relating to motif
positions.
}
\description{
\code{summit_to_motif()} calculates the distance between each motif and its
nearest summit. A peak summit represents the exact nucleotide where signal
enrichment is highest.
}
\examples{
\dontrun{
pwm <-
  read_motif_file(motif_file = "./MA0018.3.meme", file_format = "meme")

summit_to_motif(peak_file = "read1_01_peaks.narrowPeak",
                pwm = pwm,
                min_score = NULL,
                optimal_min_score = TRUE,
                seed = 123
                )
                }
}
