% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summit_to_motif.R
\name{summit_to_motif}
\alias{summit_to_motif}
\title{Calculate the distance between peak summits and motifs}
\usage{
summit_to_motif(
  peak_input,
  control_input = control_input,
  pwm = pwm,
  genome_build = genome_build,
  num_samples = num_samples,
  sample_size = sample_size,
  num_threads = num_threads,
  deoptim_lower = deoptim_lower,
  deoptim_upper = deoptim_upper,
  deoptim_np = deoptim_np,
  deoptim_itermax = deoptim_itermax,
  deoptim_strategy = deoptim_strategy,
  seed = seed
)
}
\arguments{
\item{peak_input}{Either a path to the narrowPeak file or a GRanges peak
object generated by \code{read_peak_file()}.}

\item{control_input}{Method for generating control sequences.
\code{"shuffle"} will shuffle the input sequences, whereas \code{"adjacent"}
selects real sequences immediately adjacent to each peak. The default is
\code{"shuffle"}. This argument is ignored if \code{optimal_min_score} is set
to \code{FALSE}.}

\item{pwm}{An object of class \code{PWMatrix}.}

\item{genome_build}{The genome build that the peak sequences should be
derived from.}

\item{num_samples}{Number of samples to use to calculate the optimal
min_score threshold. This option is only used if optimal_min_score is set to
TRUE.}

\item{sample_size}{Number of peaks in each sample. This option is only used
if optimal_min_score is set to TRUE.}

\item{num_threads}{Number of threads to use for min_score optimisation. We
recommend setting this equal to the number of samples for the greatest speed
up.}

\item{deoptim_lower}{The lower bound of the parameter search space.}

\item{deoptim_upper}{The upper bound of the parameter search space.}

\item{deoptim_np}{The number of population members.}

\item{deoptim_itermax}{The number of population generations the optimisation
procedure will use.}

\item{deoptim_strategy}{The differential evolution strategy the optimisation
procedure will use.}

\item{seed}{Set this to an integer if you want the same random background
sequences across function calls (given the same peak_input).}
}
\value{
GRanges peak object with metadata columns relating to motif
positions.
}
\description{
\code{summit_to_motif()} calculates the distance between each motif and its
nearest summit. A peak summit represents the exact nucleotide where signal
enrichment is highest.
}
\examples{
\dontrun{
pwm <-
  read_motif_file(motif_file = "./MA0018.3.meme", file_format = "meme")

summit_to_motif(peak_file = "read1_01_peaks.narrowPeak",
                pwm = pwm,
                min_score = NULL,
                optimal_min_score = TRUE,
                seed = 123
                )
                }
}
